Updates link to login page
Corrects the link on the registration page that directs
users to the login page. It now points to the correct
'index.php' file.

Updates login link on registration page

Corrects the link on the registration page that directs users to the login page.

It now points to the correct 'index.php' file.